fix: MyBook API findAllByUserID 에러 해결

+++ src/main/java/megabooks/megabooks/domain/myBook/dto/MyBookResponseDTO.java
@@ -14,7 +14,7 @@ public class MyBookResponseDTO {
         private String bookTitle;
         private String bookAuthor;
         private String bookPublisher;
-        private int myBookProcess;
+        private double myBookProcess;
         private String bookImgUrl;
         private BookGenre bookGenre;
         private double stars;
